Add _standbyBoxesList list Add AddBoxToStandbyList method Add PlaceBoxFromStandbyList method Add CreateStandbyBoxes method
In standby mode, boxes move to a point behind the screen and are paused; when the method for placing a box is called, it takes one of the boxes from the waiting list, places it behind the screen in a random Y coordinate, within the screen height, and unpauses it.
When the box goes off-screen, it says it went off-screen and the object controller puts it on a waiting list.

Some little rafactor project
Edit .gitignore for VSGodot folder